Improve infinite recursion detection
This detects infinite recursion within the same thread (i.e. when a
thread waits on a thunk that it's already evaluating). This is done by
storing a thread ID in pending/awaited values.

It does not detect cycles between threads, e.g. in `rec { x = y; y =
x; }` if `x` and `y` are marked as pending by different threads. But
those are much less likely in practice.