printValueAsJSON(): Don't wait for futures
Just force evaluation in the background, no need to wait for it. If
the caller gets to a value before the worker threads do, no problem -
the caller will just do the evaluation instead, which is probably
faster.

Not waiting removes a deadlock possibility, where the calling thread
waits for some work items that are themselves waiting for a thunk that
is being evaluated by the calling thread. Fixes #220.