Print the Nix version when using `-vv` (or more) verbosity
Knowing the Nix version a user is running on is useful when attempting
to debug, and if the user is giving you the output of e.g. `nix build
... --debug`, you won't have to ask for the version separately since
it'll be in the log.

Requiring 2 `v`s was a conscious decision: `-v` doesn't really
print anything that would be useful for debugging anyways, but `-vv`
might. Rather than printing more likely-unnecessary information, just
sequester it away to Talkative or higher verbosities.