Optimize version 1 nario import
`StringSink` turns out to be pretty expensive if you have a bunch of
very large strings, since once the string gets big, it's allocated
outside of the regular heap via mmap. So every 4096 bytes we get a
page fault, and the whole string is returned to the OS when
`StringSink` is destroyed. So we have a huge system time overhead.

So reuse the `StringSink` for each NAR. This speeds up importing a 15
GB NixOS system closure (with numerous large NARs) from 15.7s to 7.4s.