Backward compatibility hack for Git inputs that depend on Git filters
Before Nix 2.20, we used git, which applies Git filters (in particular
doing end-of-line conversion based on .gitattributes). In 2.20, we
switched to libgit2 and stopped applying filters, which is probably
better for reproducibility. However, that breaks existing lock files /
fetchTree calls for Git inputs that use those filters, since it
invalidates the NAR hash.

So as a backward compatibility hack, we now check the NAR hash
computed over the Git tree without filtering applied. If there is a
hash mismatch, we try again *with* filtering. If that succeeds, we
print a warning and return the filtered tree.