Backward compatibility hack for Git inputs depending on Nix < 2.20 behaviour
Before Nix 2.20, we used git, which applies Git filters (in particular
doing end-of-line conversion based on .gitattributes), `export-ignore`
and `export-subst`. In 2.20, we switched to libgit2 and stopped
applying those, which is probably better for reproducibility
(e.g. `export-subst` can do pretty crazy things). However, that breaks
existing lock files / fetchTree calls for Git inputs that rely on
them, since it invalidates the NAR hash.

So as a backward compatibility hack, we now check the NAR hash
computed over the Git tree without them. If there is a hash mismatch,
we try again *with* them (by calling `git archive`, just like Nix <
2.20). If that succeeds, we print a warning and return the filtered
tree. This is inefficient, but should be fine for legacy flakes.